requirements.txt
nicegui
anthropic
python-dotenv
